requirement.txt
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
comtypes==1.2.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.0
ipykernel==6.23.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==8.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyzmq==25.0.2
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.24.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.15.0
